+++ src/components/tools/abc-analysis/ABCAnalysis.tsx
@@ -3,9 +3,10 @@ import {Tool} from "../../../general-components/Tool/Tool";
 import {SaveResource} from "../../../general-components/Datastructures";
 
 import "./abc-analysis.scss";
+import {SteppableTool} from "../../../general-components/Tool/SteppableTool/SteppableTool";
 
 
-class ABCAnalysis extends Tool {
+class ABCAnalysis extends SteppableTool {
 
     constructor(props: any) {
         super(props);
@@ -13,7 +14,6 @@ class ABCAnalysis extends Tool {
         this.setID(4);
         this.setToolname("ABC Analyse");
         this.setToolIcon(faChartPie);
-        // Maintenance Mode
         this.setMaintenance(true);
     }
 
@@ -27,10 +27,6 @@ class ABCAnalysis extends Tool {
 
     protected renderTutorial() {
         return null;
-    }
-
-    protected renderNew() {
-        return this.getStepComponent();
     }
 
     protected renderView(tool: SaveResource) {
