+++ src/general-components/Tool/SteppableTool/StepComponent/step-component-desk.scss
@@ -1,5 +1,5 @@
-@import "src/scss/variables/variables";
-@import "~bootstrap/scss/bootstrap-utilities";
+@import "../../../../scss/variables/variables";
+@import "../../../../../node_modules/bootstrap/scss/bootstrap-utilities";
 
 @include media-breakpoint-up(xl) {
   .stepContainer > .stepTabContainer {
