+++ src/general-components/Tool/SteppableTool/StepComponent/step-component.scss
@@ -1,4 +1,4 @@
-@import "src/scss/variables/variables";
+@import "../../../../scss/variables/variables";
 
 $tabColor: #6e6e6e;
